+++ src/utils/types.ts
@@ -3,3 +3,9 @@ export type Pipelineable<T> =
   | Iterable<T>
   | IterableCreator<T>
 ;
+
+
+export function unpipeline<T>(pipelineable: Pipelineable<T>): Iterable<T> {
+  if (typeof pipelineable === 'function') return pipelineable();
+  return pipelineable;
+}