+++ changelog.md
@@ -1,5 +1,18 @@
 # Changelog
 
+## 0.0.2 - 2021-xx-xx
+
+### Added
+
+- `GetURI` utility type
+- Implement `HKT` interface
+- Implement `HKT` interface
+
+### Changed
+
+- Fixed type ambiguity in `Pipeline.prototype.flatten`
+- Fixed type ambiguity in `Pipeline.prototype.flatMap`
+
 ## 0.0.1 - 2021-08-16
 
 ### Added